project file structure rearranged enabled SocketIo
