+++ app/database/__init__.py
@@ -1,0 +1,4 @@
+from flask_mongoengine import MongoEngine
+
+# initialization of Flask-MongoEngine.
+db = MongoEngine()
