+++ app/resful/__init__.py
@@ -1,0 +1,4 @@
+from flask_restful import Api
+
+
+api = Api()
