+++ app/resful/deliveries.py
@@ -1,13 +1,11 @@
-
-import mongoengine
 from bson import ObjectId
 
 from flask_restful import Resource
-from app import socketIO
-from database.models.Delivery import Delivery
-from flask import Response, request
+from app.socketio import socketio
+from app.database.models.Delivery import Delivery
+from flask import request
 from flask_jwt_extended import jwt_required, get_current_user
-from database.models.User import User
+from app.database.models.User import User
 
 # TODO: plan end-points and resource needed
 # TODO: validate args (marshmallow?)
@@ -20,23 +18,18 @@ cancel/delete delivery - need to make sure only the user whom added the delivery
 
 class Deliveries(Resource):
 
-
     def get(self, delivery_id: str):
         delivery = Delivery.objects(id=delivery_id).first_or_404('Delivery not found').to_json()
         return delivery, 200
-
 
     @jwt_required()
     def put(self, delivery_id: str):
         req_body = request.args.to_dict()
         Delivery.objects(id=delivery_id).update(**req_body)
         if req_body['status']:
-            print (11111111)
-            socketIO.emit("delivery accepted", "sdkjrgcnd")
+            print(11111111)
+            socketio.emit("delivery accepted", "sdkjrgcnd")
         return 204
-
-
-
 
     @jwt_required()
     def delete(self, delivery_id: str):
@@ -76,4 +69,3 @@ class DeliveriesList(Resource):
         user.deliveriesRef.append(delivery)
         user.save()
         return str(delivery.id), 200
-
