+++ main.py
@@ -1,0 +1,24 @@
+from app import create_api
+from app.database.db import initialize_db
+from app.socketio import create_socketio, socketio
+from app.jwt_manger import initialize_jwt
+
+# TODO: marshmallow validation of IO
+# TODO: Generate Swagger
+
+
+# Initialise restful api
+app = create_api(debug=True)
+
+# Initialise flask-socketio
+create_socketio(app)
+
+# Initialise flask_mongoengine
+initialize_db(app)
+
+# Initialise flask_jwt_extended
+initialize_jwt(app)
+
+
+if __name__ == '__main__':
+    socketio.run(app, host='0.0.0.0')
